fix style check error
