+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/optim/EvaluatorSpec.scala
@@ -87,58 +87,4 @@ class EvaluatorSpec extends SparkContextLifeCycle with Matchers {
     result(1)._1.result()._1 should be (1f)
     result(2)._1.result()._1 should be (2.3044279f+-0.000001f)
   }
-
-  "Evaluator with dnn backend" should "be correct" in {
-    Engine.setEngineType(MklDnn)
-    RNG.setSeed(100)
-    val tmp = new Array[Sample[Float]](100)
-    var i = 0
-    while (i < tmp.length) {
-      val input = Tensor[Float](28, 28).fill(0.8f)
-      val label = Tensor[Float](1).fill(1.0f)
-      tmp(i) = Sample(input, label)
-      i += 1
-    }
-    val model = LeNet5(classNum = 10)
-    val dataSet = DataSet.array(tmp, sc).toDistributed().data(train = false).repartition(4)
-
-    val result = model.evaluate(dataSet, Array(new Top1Accuracy[Float](), new Top5Accuracy[Float](),
-      new Loss[Float](CrossEntropyCriterion[Float]())))
-
-    dataSet.getNumPartitions should be(4)
-    result(0)._1 should be (new AccuracyResult(0, 100))
-    result(1)._1 should be (new AccuracyResult(100, 100))
-    result(2)._1 should be (new LossResult(16.130993f, 7))
-    result(0)._1.result()._1 should be (0f)
-    result(1)._1.result()._1 should be (1f)
-    result(2)._1.result()._1 should be (2.3044279f+-0.000001f)
-    Engine.setEngineType(MklBlas)
-  }
-
-  "Evaluator MiniBatch with dnn backend" should "be correct" in {
-    Engine.setEngineType(MklDnn)
-    RNG.setSeed(100)
-    val tmp = new Array[MiniBatch[Float]](25)
-    var i = 0
-    while (i < tmp.length) {
-      val input = Tensor[Float](4, 28, 28).fill(0.8f)
-      val label = Tensor[Float](4).fill(1.0f)
-      tmp(i) = MiniBatch(input, label)
-      i += 1
-    }
-    val model = LeNet5(classNum = 10)
-    val dataSet = DataSet.array(tmp, sc).toDistributed().data(train = false).repartition(4)
-
-    val result = model.evaluate(dataSet, Array(new Top1Accuracy[Float](), new Top5Accuracy[Float](),
-      new Loss[Float](CrossEntropyCriterion[Float]())))
-
-    dataSet.getNumPartitions should be(4)
-    result(0)._1 should be (new AccuracyResult(0, 100))
-    result(1)._1 should be (new AccuracyResult(100, 100))
-    result(2)._1 should be (new LossResult(57.610695f, 25))
-    result(0)._1.result()._1 should be (0f)
-    result(1)._1.result()._1 should be (1f)
-    result(2)._1.result()._1 should be (2.3044279f+-0.000001f)
-    Engine.setEngineType(MklBlas)
-  }
 }
